+++ packages/components/common/social-icons/src/social-icons.tsx
@@ -1,5 +1,5 @@
 import { useStackShiftUIComponents } from "@stackshift-ui/system";
-import cn from "classnames";
+import { cn } from "@stackshift-ui/system";
 import type { ElementType, HTMLProps, ReactNode } from "react";
 import {
   FaDiscord,