+++ packages/components/common/swiper-pagination/src/swiper-pagination.tsx
@@ -1,5 +1,5 @@
 import { DefaultComponent, useStackShiftUIComponents } from "@stackshift-ui/system";
-import cn from "classnames";
+import { cn } from "@stackshift-ui/system";
 import type { ElementType, HTMLProps, ReactNode } from "react";
 
 type ColorScheme = "blue" | "white";