+++ packages/components/layout/grid-item/src/grid-item.tsx
@@ -1,5 +1,5 @@
 import { DefaultComponent, useStackShiftUIComponents } from "@stackshift-ui/system";
-import cn from "classnames";
+import { cn } from "@stackshift-ui/system";
 import type { ElementType, HTMLProps, ReactNode } from "react";
 
 export interface GridItemProps extends Omit<HTMLProps<HTMLElement>, "as" | "span"> {