+++ packages/components/layout/section/src/section.tsx
@@ -1,5 +1,5 @@
 import { DefaultComponent, useStackShiftUIComponents } from "@stackshift-ui/system";
-import cn from "classnames";
+import { cn } from "@stackshift-ui/system";
 import type { ElementType, HTMLProps, ReactNode } from "react";
 
 type Width = "sm" | "md" | "lg" | "xl" | "2xl" | "full" | number;