+++ packages/components/common/avatar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackshift-ui/avatar
 
+## 6.1.1-fixes.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackshift-ui/flex@6.1.1-fixes.0
+
 ## 6.1.0
 
 ### Major Changes