+++ packages/components/common/stats-card/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackshift-ui/stats-card
 
+## 6.1.1-fixes.0
+
+### Patch Changes
+
+- Changin of import cn from 'classnames' to import cn from @stackshift-ui/system
+- Updated dependencies
+  - @stackshift-ui/text@6.1.1-fixes.0
+
 ## 6.1.0
 
 ### Major Changes