+++ packages/components/sections/footer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackshift-ui/footer
 
+## 6.1.1-fixes.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackshift-ui/social-icons@6.1.1-fixes.0
+  - @stackshift-ui/container@6.1.1-fixes.0
+  - @stackshift-ui/grid-item@6.1.1-fixes.0
+  - @stackshift-ui/section@6.1.1-fixes.0
+  - @stackshift-ui/text@6.1.1-fixes.0
+  - @stackshift-ui/flex@6.1.1-fixes.0
+  - @stackshift-ui/grid@6.1.1-fixes.0
+
 ## 6.1.0
 
 ### Major Changes