+++ packages/components/sections/logo-cloud/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @stackshift-ui/logo-cloud
 
+## 6.1.1-fixes.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackshift-ui/container@6.1.1-fixes.0
+  - @stackshift-ui/heading@6.1.1-fixes.0
+  - @stackshift-ui/section@6.1.1-fixes.0
+  - @stackshift-ui/text@6.1.1-fixes.0
+  - @stackshift-ui/flex@6.1.1-fixes.0
+
 ## 6.1.0
 
 ### Major Changes