+++ packages/components/sections/testimonial/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackshift-ui/testimonial
 
+## 6.1.1-fixes.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackshift-ui/swiper-pagination@6.1.1-fixes.0
+  - @stackshift-ui/swiper-button@6.1.1-fixes.0
+  - @stackshift-ui/container@6.1.1-fixes.0
+  - @stackshift-ui/heading@6.1.1-fixes.0
+  - @stackshift-ui/section@6.1.1-fixes.0
+  - @stackshift-ui/text@6.1.1-fixes.0
+  - @stackshift-ui/flex@6.1.1-fixes.0
+  - @stackshift-ui/avatar@6.1.1-fixes.0
+
 ## 6.1.0
 
 ### Major Changes