+++ packages/core/react/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @stackshift-ui/react
 
+## 7.0.1-fixes.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackshift-ui/swiper-pagination@6.1.1-fixes.0
+  - @stackshift-ui/checkbox-group@6.1.1-fixes.0
+  - @stackshift-ui/swiper-button@6.1.1-fixes.0
+  - @stackshift-ui/social-icons@6.1.1-fixes.0
+  - @stackshift-ui/navigation@6.1.1-fixes.0
+  - @stackshift-ui/input-file@6.1.1-fixes.0
+  - @stackshift-ui/stats-card@6.1.1-fixes.0
+  - @stackshift-ui/container@6.1.1-fixes.0
+  - @stackshift-ui/grid-item@6.1.1-fixes.0
+  - @stackshift-ui/heading@6.1.1-fixes.0
+  - @stackshift-ui/section@6.1.1-fixes.0
+  - @stackshift-ui/radio@6.1.1-fixes.0
+  - @stackshift-ui/text@6.1.1-fixes.0
+  - @stackshift-ui/flex@6.1.1-fixes.0
+  - @stackshift-ui/grid@6.1.1-fixes.0
+  - @stackshift-ui/header@6.1.1-fixes.0
+  - @stackshift-ui/testimonial@6.1.1-fixes.0
+  - @stackshift-ui/app-promo@6.1.1-fixes.0
+  - @stackshift-ui/contact@6.1.1-fixes.0
+  - @stackshift-ui/footer@6.1.1-fixes.0
+  - @stackshift-ui/blog@6.1.1-fixes.0
+  - @stackshift-ui/call-to-action@6.1.1-fixes.0
+  - @stackshift-ui/faqs@6.1.1-fixes.0
+  - @stackshift-ui/features@6.1.1-fixes.0
+  - @stackshift-ui/how-it-works@6.1.1-fixes.0
+  - @stackshift-ui/logo-cloud@6.1.1-fixes.0
+  - @stackshift-ui/newsletter@6.1.1-fixes.0
+  - @stackshift-ui/portfolio@6.1.1-fixes.0
+  - @stackshift-ui/signin-signup@6.1.1-fixes.0
+  - @stackshift-ui/statistics@6.1.1-fixes.0
+  - @stackshift-ui/team@6.1.1-fixes.0
+  - @stackshift-ui/text-component@6.1.1-fixes.0
+  - @stackshift-ui/blockstyle@6.1.1-fixes.0
+  - @stackshift-ui/avatar@6.1.1-fixes.0
+  - @stackshift-ui/badge@6.1.1-fixes.0
+
 ## 7.0.0
 
 ### Major Changes